clean id handling

+++ x/authenticator/post/post.go
@@ -68,23 +68,9 @@ func (ad AuthenticatorDecorator) PostHandle(
 			return sdk.Context{}, errorsmod.Wrap(sdkerrors.ErrUnauthorized, fmt.Sprintf("failed to get authentication data for message %d", msgIndex))
 		}
 		for _, accountAuthenticator := range accountAuthenticators { // This should execute on the authenticators used to authenticate the msg
-			// TODO: defence in depth checks on lengths everywhere? or maybe a better datastructure for this?
-			if usedAuthenticators[msgIndex] != int64(accountAuthenticator.Id) {
+			if usedAuthenticators[msgIndex] != accountAuthenticator.Id {
 				continue
 			}
-
-			// TODO: I think this can go away now that we only execute post on used authenticators.
-			//  B ut will need to reason a bit more about this before fully removing
-			//
-			//// We want to skip `ConfirmExecution` if the authenticator is newly added
-			//// since Authenticate & Track are called on antehandler but newly added authenticator
-			//// so Authenticate & Track on newly added authenticator has not been called yet
-			//// which means the authenticator is not ready to confirm execution
-			//if !accountAuthenticator.IsReady {
-			//	key := string(types.KeyAccountId(account, accountAuthenticator.Id))
-			//	nonReadyAccountAuthenticatorKeys[key] = struct{}{}
-			//	continue
-			//}
 
 			authenticator := accountAuthenticator.AsAuthenticator(ad.authenticatorKeeper.AuthenticatorManager)
